+++ NEWS.md
@@ -16,7 +16,6 @@ the number of dimensions to reach a target *pseudo R squared*
 - in `ggplotSampleMDS()`, added `flipXAxis`, `flipYAxis` 
 to possibly ease low dimensional projection comparisons
 - in `ggplotSampleMDS()`, added `displayArrowLabels` to discard
-the arrow labels in biplot, while keeping the ability to interactively show
-them when using ggplot object in `ggplotly()`. Also added `arrowThreshold`.
+the arrow labels in biplot. Also added `arrowThreshold`.
 Moved arrow labels toward the end of the arrows.