+++ NEWS.md
@@ -20,3 +20,9 @@ the arrow labels in biplot. Also added `arrowThreshold`.
 Moved arrow labels toward the end of the arrows.
 - in `ggplotSampleMDS()` and `ggplotSampleMDSShepard()`: added 
 `displayPseudoRSq` parameter.
+
+### CytoMDS 0.99.4
+- in `getPairwiseEMDDist()`, added a second flowSet argument. When the two
+flowSet arguments are non-null, distances are calculated for all sample pairs, 
+where the first element comes from `fs`, 
+and the second element comes from `fs2`.